国赛/src/Ai_main/requirements.txt
facenet_pytorch==2.5.3
ipython==8.29.0
matplotlib==3.5.3
numpy==1.21.5
opencv_python==4.10.0.84
pandas==1.3.5
Pillow==11.0.0
scikit_learn==1.0.2
scipy==1.7.3
tensorboardX==2.6.2.2
torch==1.13.1
torchvision==0.14.1
